Add files via upload
I have created a script for failover between two nms instances. As nim does not have HA approach, we tried created a script that do a pseudo-HA, where the user should activate the script in order to have a complete failover method.

This bash script is designed to be executed in a scenario where the Keepalived service is running on two servers in a High Availability (HA) configuration. Keepalived is a tool commonly used for setting up high availability for Linux systems, and it works by allowing multiple servers to share an IP address and ensure that one server takes over in case the other fails. The script is structured to run on the server currently in the MASTER state according to Keepalived.
1.	Define IP Address: The script starts by defining the IP address of the other server (OTHER_SERVER_IP).
2.	Check Keepalived State: It checks whether the current server is in the MASTER state by examining the output of the ip addr show command and looking for a specific IP address (10.150.209.138). If the server is the MASTER, the script proceeds with its execution; otherwise, it exits.
3.	Downgrade and Upgrade Priority: It downgrades the priority on the current MASTER server and upgrades the priority on the other server. This is a mechanism in Keepalived to determine which server should be the MASTER in case of a failover.
4.	Restart Keepalived on the Other Server: It restarts the Keepalived service on the other server to apply the priority changes.
5.	Sync Password File: It uses rsync to synchronize the password file (/etc/nms/nginx/.htpasswd) between the servers.
6.	Restart Keepalived on the Current Master: It restarts the Keepalived service on the current MASTER server.
7.	Switch Roles: It prints a message indicating that Keepalived roles have switched successfully.
8.	Send Latest File to the Other Server: It identifies the latest file in /tmp/ starting with "nms" and sends it to the other server using scp. We also have added a crontab which create once a week a backup file for the nim.
9.	Stop Services and Run Restore Script: It stops various services on the other server and runs a restore script on it, assuming the presence of a file in /tmp/ starting with "nms". The script then restarts the services.
10.	Completion Message: It prints messages indicating the completion of the script execution.